updated User model and controller

+++ server/controllers/usersController.js
@@ -4,27 +4,43 @@ const asyncHandler = require("express-async-handler");
 const bcrypt = require("bcrypt");
 const isBase64 = require("is-base64");
 
-// @desc Get user data
+// @desc Login user
 // @route GET /user
 // @access Private
 const getUser = asyncHandler(async (req, res) => {
-    const { id } = req.body;
+    const { username, email, password: pass } = req.body;
 
     // Confirm data
-    if (!id) {
-        return res.status(400).json({ message: "User ID is required" });
+    if (!((username && pass) || (email && pass))) {
+        return res.status(400).json({ message: "All fields are required" });
     }
 
     // Check if user exists
-    const user = await User.findById(id).select("-password").lean();
+    let user;
+    if (username) {
+        user = await User.findOne({ username }).lean();
+    } else if (email) {
+        user = await User.findOne({ email }).lean();
+    }
     if (!user) {
         return res.status(404).json({ message: "User not found" });
     }
 
-    res.json(user);
+    // Check if password is correct
+    const passwordEquals = await bcrypt.compare(pass, user.password);
+    if (!passwordEquals) {
+        return res.status(401).json({
+            message: "Password is invalid",
+        });
+    }
+
+    // Remove password and document version
+    const { password, __v, ...data } = user;
+
+    res.json(data);
 });
 
-// @desc Create new user
+// @desc Create Register user
 // @route POST /user
 // @access Public
 const createNewUser = asyncHandler(async (req, res) => {
@@ -44,9 +60,30 @@ const createNewUser = asyncHandler(async (req, res) => {
         return usernamePattern.test(username);
     };
     if (!isValidUsername(username)) {
-        return res.status(409).json({
+        return res.status(401).json({
             message:
                 "Username is invalid. It should contain only letters, numbers, and underscores",
+        });
+    }
+
+    // Check if email valid
+    const isValidEmail = (email) => {
+        // Define a regular expression pattern for valid emails
+        const usernamePattern = /^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,4}$/i;
+
+        // Test the email against the pattern
+        return usernamePattern.test(email);
+    };
+    if (!isValidEmail(email)) {
+        return res.status(401).json({
+            message: "Email is invalid",
+        });
+    }
+
+    // Check if password valid
+    if (password.length < 8) {
+        return res.status(400).json({
+            message: "The password needs to be at least 8 characters long",
         });
     }
 
@@ -126,10 +163,10 @@ const updateUser = asyncHandler(async (req, res) => {
 // @route DELETE /user
 // @access Private
 const deleteUser = asyncHandler(async (req, res) => {
-    const { id } = req.body;
+    const { id, password: pass } = req.body;
 
     // Confirm data
-    if (!id) {
+    if (!id || !pass) {
         return res.status(400).json({ message: "User ID Required" });
     }
 
@@ -139,8 +176,16 @@ const deleteUser = asyncHandler(async (req, res) => {
         return res.status(404).json({ message: "User not found" });
     }
 
+    // Check if password is correct
+    const passwordEquals = await bcrypt.compare(pass, user.password);
+    if (!passwordEquals) {
+        return res.status(401).json({
+            message: "Password is invalid",
+        });
+    }
+
     // Delete collections of user
-    const deletedCollections = await Collection.deleteMany({
+    await Collection.deleteMany({
         owner: id,
     });
 
@@ -149,7 +194,6 @@ const deleteUser = asyncHandler(async (req, res) => {
 
     res.json({
         message: `User ${deletedUser.username} deleted`,
-        deletedCollections,
     });
 });
